+++ src/main/java/sic/modelo/RenglonPedido.java
@@ -1,6 +1,7 @@
 package sic.modelo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 
@@ -10,8 +11,8 @@ public class RenglonPedido implements Serializable {
 
     private long id_RenglonPedido;
     private Producto producto;
-    private double cantidad;
-    private double descuento_porcentaje;
-    private double descuento_neto;
-    private double subTotal;
+    private BigDecimal cantidad;
+    private BigDecimal descuento_porcentaje;
+    private BigDecimal descuento_neto;
+    private BigDecimal subTotal;
 }
